portfolio backend started

+++ app/src/main/java/com/example/kind/model/Portfolio.kt
@@ -2,7 +2,7 @@ package com.example.kind.model
 
 data class Portfolio(
     val organization : String,
-    val pct : Float,
-    val spend : Float,
-    val total : Float,
+    val pct : Double,
+    val spend : Double,
+    val total : Double,
 )
